캠페인 리스트 맨 아래가 잘리는 현상 수정

+++ app/src/main/java/com/example/bigwalkclone/model/CampaignModel.kt
@@ -17,7 +17,7 @@ data class CampaignModel(
     @SerializedName("detailThumbnailImagePath")
     val detailThumbnailImagePath: String,
     @SerializedName("donatedSteps")
-    val donatedSteps: Int,
+    val donatedSteps: Long,
     @SerializedName("enabled")
     val enabled: Boolean,
     @SerializedName("endDate")
@@ -27,7 +27,7 @@ data class CampaignModel(
     @SerializedName("formattedStartDate")
     val formattedStartDate: String,
     @SerializedName("goalPoint")
-    val goalPoint: Int,
+    val goalPoint: Long,
     @SerializedName("id")
     val id: Int,
     @SerializedName("largeListThumbnailImagePath")
